requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pymongo[srv]
strawberry-graphql[fastapi]
